+++ packages/arithmico-api/src/modules/security-policy/commands/delete-security-policy/delete-security-policy.controller.ts
