+++ packages/arithmico-api/src/modules/security-policy/commands/rename-security-policy/rename-security-policy.command.ts
