+++ packages/arithmico-api/src/modules/security-policy/commands/set-policy-attributes/set-security-policy-attributes.request.body.dto.ts
