+++ packages/arithmico-api/src/modules/security-policy/common/security-policy.dto.ts
