+++ packages/arithmico-api/src/modules/security-policy/queries/get-available-security-attributes/get-available-security-attributes.query.ts
