+++ packages/arithmico-api/src/modules/security-policy/queries/get-security-policies/get-security-policies.response.dto.ts
