+++ packages/arithmico-api/src/modules/security-policy/queries/get-security-policy-by-id/get-security-policy-by-id.controller.ts
