+++ packages/arithmico-api/src/modules/user-group/queries/get-security-policies-attached-to-user-group/get-security-policies-attached-to-user-group.response.dto.ts
@@ -1,3 +1,3 @@
-import { SecurityPolicyDto } from '../../../policy-policy/common/security-policy.dto';
+import { SecurityPolicyDto } from '../../../security-policy/common/security-policy.dto';
 
 export class GetSecurityPoliciesAttachedToUserGroupResponseDto extends SecurityPolicyDto {}
