+++ packages/arithmico-api/src/modules/user/queries/get-security-policies-attached-to-user/get-security-policies-attached-to-user.response.dto.ts
@@ -1,3 +1,3 @@
-import { SecurityPolicyDto } from '../../../policy-policy/common/security-policy.dto';
+import { SecurityPolicyDto } from '../../../security-policy/common/security-policy.dto';
 
 export class GetSecurityPoliciesAttachedToUserResponseDto extends SecurityPolicyDto {}
